requirements.txt
git+git://github.com/django/django.git@1.6b2
Pillow==2.0.0
PyYAML==3.10
South==0.8.1
Unidecode==0.04.13
beautifulsoup4==4.2.1
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.28dev-r0
django-braces==1.2.2
-e git+https://github.com/marky1991/django-generic-m2m.git@c9db92ff32ddd98ec0ff01027d4b9ca808e5b5a0#egg=django_generic_m2m-dev
fuzzywuzzy==0.2
gevent==1.0dev
html5lib==1.0b2
lxml==3.2.1
simplejson==3.3.0
virtualenv==1.10.1